add grass shader experimentation and test scene/assets (WIP)
I'm checking out one of the geometry shader tutorials from the trello board (the one by Chen JD). It's a shader using the Built-in Render Pipeline though, but it appears to work! It just probably doesn't take advantage of HDRP features?

I'm studying the example code, so far I've only partially implemented it - but you can press the play button in Test-grass_and_foliage to see the progress.